requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
elevenlabs==1.7.0
exceptiongroup==1.2.2
executing==2.0.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-text-splitters==0.2.2
langsmith==0.1.104
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.10.7
packaging==24.1
parso==0.8.4
platformdirs==4.2.2
# Editable install with no version control (PROJECT==0.0.0)

prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
semantic-version==2.10.0
setuptools-rust==1.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
torch==1.10.1
torchaudio==2.2.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
websockets==13.0
Werkzeug==3.0.4
yarl==1.9.4
zipp==3.20.0

-e .